pyproject.toml
[tool.poetry]
name = "drf-tutorial-lecture"
version = "0.1.0"
description = ""
authors = ["johnkdo <johnkdo@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
pygments = "^2.6.1"
django = "<3.0"
notebook = "^6.0.3"
django-extensions = "^2.2.9"
djangorestframework = "^3.11.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
